// shared/apps/weather/weatherStates.d.ts
declare enum weatherStates {
    clear = 0,
    partlyCloudy = 1,
    cloudy = 2,
    fog = 3,
    lightRain = 4,
    rain = 5,
    heavyRain = 6,
    lightSnow = 7,
    snow = 8,
    heavySnow = 9,
    lightRainShowers = 10,
    rainShowers = 11,
    heavyRainShowers = 12,
    thunder = 13
}
export { weatherStates };